# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Pariraksakah Backend API
  version: 1.0.0
  description: |
    Swagger/OpenAPI documentation for backend APIs exposed by the API Gateway.

    Notes:
    - Public endpoints are available without JWT.
    - Protected endpoints require Bearer JWT unless otherwise stated.
    - Several protected routes are proxy pass-through routes to internal services.
servers:
  - url: http://localhost:8080
    description: Local API Gateway
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login (public gateway route)
      description: Proxied to access-control service auth login.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Login response from access-control
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/v1/auth/verify:
    post:
      tags: [Auth]
      summary: Verify token (public gateway route)
      description: Proxied to access-control service auth verify.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Verification response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Legacy login route
      description: Public access-control proxy route.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Auth response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /auth/token:
    post:
      tags: [Auth]
      summary: Legacy token route
      description: Public access-control proxy route.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Token response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RolloutToken:
      type: apiKey
      in: header
      name: X-Rollout-Token
...